+++ src/screens/TransactionRequests/TransactionHistoryScreen.tsx
@@ -1,10 +1,47 @@
 import * as React from 'react';
-import { View, Text } from 'react-native';
+import { View, Text, FlatList } from 'react-native';
+import RequestCardComponent from '../../components/RequestCardComponent';
+
+const DATA = [
+  {
+    id: "bd7acbea-c1b1-46c2-aed5-3ad53abb28ba",
+    title: "First Item",
+  },
+  {
+    id: "3ac68afc-c605-48d3-a4f8-fbd91aa97f63",
+    title: "Second Item",
+  },
+  {
+    id: "58694a0f-3da1-471f-bd96-145571e29d72",
+    title: "Third Item",
+  },
+];
 
 export default function TransactionHistoryScreen() {
     return (
-      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-        <Text>Transaction History Screen</Text>
-      </View>
+      <View
+      style={{
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      
+      <FlatList
+            data={DATA}
+            renderItem={({ item }) => (
+              <RequestCardComponent
+                // _id={item._id}
+                // amount={item.amount}
+                // stars={item.stars}
+                // rating={item.rating}
+                // type={item.type}
+                // transaction={item}
+                // deleteItem={removeDepositRequestItem}
+              />
+            )}
+            keyExtractor={(item) => item.id}
+          />
+    </View>
     );
   }
